backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
psycopg2-binary